ENH Split tests into module-specific files; add shared fixtures and helpers
- Replace tests/test_diffractometer.py with six module-specific files:
  test_axes.py, test_rotation.py, test_stage.py, test_geometry.py,
  test_factories.py, test_display.py (310 tests, all passing)
- tests/helpers.py: importable module with Rx/Ry/Rz reference rotation
  matrices and STANDARD_BASIS; used in parametrize() decorators which
  evaluate at collection time before fixture injection
- tests/conftest.py: pytest fixtures shared across all test files:
  psic_geom (fresh psic() per test), reset_precision (autouse; restores
  display precision after every test)
- pyproject.toml: add pythonpath = ["tests"] so helpers.py is importable
- .github/workflows/tests.yml: add Python 3.14-dev to matrix as an
  allowed failure (continue-on-error) for early breakage warning
- AGENTS.md: roadmap checkboxes belong on the feature branch, not a
  separate post-merge commit to main
- docs/roadmap.md: check boxes for 1.4 (test split) and 1.5 (CI workflow
  update); tests.yml note updated for 3.14-dev

Contributed by: OpenCode (argo/claudesonnet46)
