+++ .github/workflows/tests.yml
@@ -15,6 +15,12 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ["3.10", "3.11", "3.12", "3.13"]
+        include:
+          # 3.14 is pre-release; run it for early warning but don't fail CI
+          - python-version: "3.14-dev"
+            continue-on-error: true
+
+    continue-on-error: ${{ matrix.continue-on-error == true }}
 
     steps:
       - uses: actions/checkout@v4
