+++ pyproject.toml
@@ -34,6 +34,7 @@ force_single_line = true
 known_first_party = ["ad_hoc_diffractometer"]
 
 [tool.pytest.ini_options]
+pythonpath = ["tests"]
 testpaths = ["tests"]
